requirements.txt
# Core dependencies (100% gratuit avec Ollama)
streamlit
langchain
langchain-community
langchain-chroma
langchain-text-splitters

# Web fetching pour documentation
requests
beautifulsoup4

# Vector database
chromadb

# Utilities
python-dotenv
